refactor: modularize crypto components
